insert のときだけ消すように

+++ perl/lib/Isuda/Web.pm
@@ -193,16 +193,23 @@ post '/keyword' => [qw/set_name authenticate/] => sub {
         ON DUPLICATE KEY UPDATE
         author_id = ?, keyword = ?, description = ?, updated_at = NOW()
     ], ($user_id, $keyword, $description) x 2);
-    $cache->delete($CACHE_KEY_KEYWORDS);
+
     $cache->delete($CACHE_KEY_HTML . ":$keyword");
-    my $entries = $self->dbh->select_all(qq[
-        SELECT keyword FROM entry WHERE description LIKE "%$keyword%"
-    ]);
 
-    for my $entry (@$entries) {
-       $cache->delete($CACHE_KEY_HTML . ":$entry->{keyword}");
+    my $last_insert_id = $self->dbh->last_insert_id;
+    # inesrt
+    if ($last_insert_id) {
+        $cache->delete($CACHE_KEY_KEYWORDS);
+        my $entries = $self->dbh->select_all(qq[
+           SELECT keyword FROM entry WHERE description LIKE "%$keyword%"
+        ]);
+
+        my @cache_keys;
+        for my $entry (@$entries) {
+            push @cache_keys, $CACHE_KEY_HTML . ":$entry->{keyword}";
+        }
+        $cache->delete_multi(\@cache_keys);
     }
-
     $c->redirect('/');
 };
